// build/ui/src/common/types.ts
interface DirectoryItemBasic {
  name: string;
  whitelisted: boolean;
  isFeatured: boolean;
}
export interface DirectoryItemOk extends DirectoryItemBasic {
  status: "ok";
  description: string; // = metadata.shortDescription || metadata.description
  avatarUrl: string; // Must be URL to a resource in a DAPPMANAGER API
  isInstalled: boolean; // Show "UPDATE"
  isUpdated: boolean; // Show "UPDATED"
  featuredStyle?: {
    featuredBackground?: string;
    featuredColor?: string;
    featuredAvatarFilter?: string;
  };
  categories: string[];
}
export interface DirectoryItemLoading extends DirectoryItemBasic {
  status: "loading";
  message?: string;
}
export interface DirectoryItemError extends DirectoryItemBasic {
  status: "error";
  message: string;
}
export type DirectoryItem =
  | DirectoryItemOk
  | DirectoryItemLoading
  | DirectoryItemError;

export interface RequestStatus {
  loading?: boolean;
  error?: string;
  success?: boolean;
}

export interface SetupWizard {
  version: "2";
  fields: SetupWizardField[];
}

export interface SetupWizardField {
  id: string;
  target?: UserSettingTarget; // Allow form questions
  // UI
  title: string;
  description: string;
  secret?: boolean;
  // Validation options
  pattern?: string;
  patternErrorMessage?: string;
  enum?: string[];
  required?: boolean;
  if?: SetupSchema | { [id: string]: SetupSchema };
}

export type UserSettingTarget =
  | { type: "environment"; name: string }
  | { type: "portMapping"; containerPort: string }
  | { type: "namedVolumeMountpoint"; volumeName: string }
  | { type: "allNamedVolumesMountpoint" }
  | { type: "fileUpload"; path: string };

export interface SetupWizardAllDnps {
  [dnpName: string]: SetupWizard;
}

export interface SetupTarget {
  [propId: string]: UserSettingTarget;
}

export interface SetupSchemaAllDnps {
  [dnpName: string]: SetupSchema;
}

export interface SetupTargetAllDnps {
  [dnpName: string]: SetupTarget;
}

export interface SetupUiJsonAllDnps {
  [dnpName: string]: SetupUiJson;
}

// Settings must include the previous user settings

export interface UserSettings {
  environment?: { [envName: string]: string }; // Env value
  portMappings?: { [containerPortAndType: string]: string }; // Host port
  namedVolumeMountpoints?: { [volumeName: string]: string }; // Host absolute path to mountpoint
  allNamedVolumeMountpoint?: string; // mountpoint
  fileUploads?: { [containerPath: string]: string }; // dataURL
  domainAlias?: string[]; // ["fullnode", "my-custom-name"]
}
// "bitcoin.dnp.dappnode.eth": {
//   environment: { MODE: "VALUE_SET_BEFORE" }
//   portMappings: { "8443": "8443"; "8443/udp": "8443" },
//   namedVolumeMountpoints: { data: "" }
//   fileUploads: { "/usr/src/app/config.json": "data:text/plain;base64,SGVsbG8sIFdvcmxkIQ%3D%3D" }
// };
export interface UserSettingsAllDnps {
  [dnpName: string]: UserSettings;
}

export interface DisclaimerAllDnps {
  [dnpName: string]: string;
}

export interface CompatibleDnps {
  [dnpName: string]: { from?: string; to: string };
  // "bitcoin.dnp.dappnode.eth": { from: "0.2.5"; to: "0.2.6" };
  // "ln.dnp.dappnode.eth": { from: null; to: "0.2.2" };
}

export interface RequestedDnp {
  name: string; // "bitcoin.dnp.dappnode.eth"
  reqVersion: string; // origin or semver: "/ipfs/Qm611" | "0.2.3"
  semVersion: string; // Always a semver: "0.2.3"
  origin?: string; // "/ipfs/Qm"
  avatarUrl: string; // "http://dappmanager.dappnode/avatar/Qm7763518d4";
  // Setup
  setupWizard?: SetupWizardAllDnps;
  settings: UserSettingsAllDnps; // MUST include the previous user settings
  // Additional data
  imageSize: number;
  isUpdated: boolean;
  isInstalled: boolean;
  // Decoupled metadata
  metadata: PackageReleaseMetadata;
  specialPermissions: SpecialPermission[];
  // Request status and dependencies
  request: {
    compatible: {
      requiresCoreUpdate: boolean;
      resolving: boolean;
      isCompatible: boolean; // false;
      error: string; // "LN requires incompatible dependency";
      dnps: CompatibleDnps;
    };
    available: {
      isAvailable: boolean; // false;
      message: string; // "LN image not available";
    };
  };
}

// Installing types

export interface ProgressLogs {
  [dnpName: string]: string;
}

export interface ProgressLogsByDnp {
  [dnpName: string]: ProgressLogs;
}

/**
 * ==============
 * ==============
 * DAPPMANAGER
 * ==============
 * ==============
 */

export type PortProtocol = "UDP" | "TCP";

interface BasicPortMapping {
  host?: number;
  container: number;
  protocol: PortProtocol;
}

export interface PortMapping extends BasicPortMapping {
  ephemeral?: boolean;
  ip?: string;
  deletable?: boolean;
}

export interface PackagePort {
  portNumber: number;
  protocol: PortProtocol;
}

interface BasicVolumeMapping {
  host: string; // path
  container: string; // dest
  name?: string;
}

export interface VolumeMapping extends BasicVolumeMapping {
  users?: string[];
  owner?: string;
  isOwner?: boolean;
  size?: number;
}

export interface Dependencies {
  [dependencyName: string]: string;
}

export type ContainerStatus =
  | "created" // created A container that has been created(e.g.with docker create) but not started
  | "restarting" // restarting A container that is in the process of being restarted
  | "running" // running A currently running container
  | "paused" // paused A container whose processes have been paused
  | "exited" // exited A container that ran and completed("stopped" in other contexts, although a created container is technically also "stopped")
  | "dead"; // dead A container that the daemon tried and failed to stop(usually due to a busy device or resource used by the container)

export type ChainDriver = "bitcoin" | "ethereum" | "monero";

export interface PackageContainer {
  id: string;
  packageName: string;
  version: string;
  isDnp: boolean;
  isCore: boolean;
  created: number;
  image: string;
  name: string;
  shortName: string;
  ip?: string; // IP of the DNP in the dappnode network
  state: ContainerStatus;
  running: boolean;
  envs?: PackageEnvs;
  ports: PortMapping[];
  volumes: VolumeMapping[];
  defaultEnvironment: PackageEnvs;
  defaultPorts: PortMapping[];
  defaultVolumes: VolumeMapping[];
  dependencies: Dependencies;
  avatarUrl: string;
  origin?: string;
  chain?: ChainDriver;
  domainAlias?: string[];
  canBeFullnode?: boolean;
  // ### TODO: Move to PackageDetails, note it will require significant
  // changes to the ADMIN UI in parts the code is not yet typed
  manifest?: Manifest;
  gettingStarted?: string;
  gettingStartedShow?: boolean;
}

export interface PackageEnvs {
  [envName: string]: string;
}

export interface PackageDetailData {
  volumes?: {
    // volumeName = bitcoin_data
    [volumeName: string]: {
      size?: string; // "823203"
      devicePath: string; // "/dev1/data/dappnode-volumes/bitcoin.dnp.dappnode.eth/data"
      mountpoint?: string; // "/dev1/data"
    };
  };
  setupWizard?: SetupWizard;
  userSettings?: UserSettings;
}

export interface ManifestUpdateAlert {
  from: string;
  to: string;
  message: string;
}

interface ManifestImage {
  hash: string;
  size: number;
  path: string;
  volumes?: string[];
  external_vol?: string[];
  ports?: string[];
  environment?: string[];
  restart?: string;
  privileged?: boolean;
  cap_add?: string[];
  cap_drop?: string[];
  devices?: string[];
  subnet?: string;
  ipv4_address?: string;
  network_mode?: string;
  command?: string;
  labels?: string[];
}
export interface Manifest extends PackageReleaseMetadata {
  name: string;
  version: string;
  avatar?: string;
}
// ...
export interface PackageBackup {
  name: string;
  path: string;
}
// ...
export interface DependencyListItem {
  name: string;
  from?: string;
  to: string;
  warningOnInstall?: string;
}

export interface CoreUpdateData {
  available: boolean;
  type?: string;
  packages: DependencyListItem[];
  changelog: string;
  updateAlerts: ManifestUpdateAlert[];
  versionId: string;
}
// ...
export interface SpecialPermission {
  name: string; // "Short description",
  details: string; // "Long description of the capabilitites"
}
// ...
export interface PackageReleaseMetadata {
  name: string;
  version: string;
  upstreamVersion?: string;
  shortDescription?: string;
  description?: string;
  type?: string;
  chain?: string;
  dependencies?: Dependencies;

  // Safety properties to solve problematic updates
  runOrder?: string[];
  restartCommand?: string;
  restartLaunchCommand?: string;

  requirements?: {
    minimumDappnodeVersion: string;
  };
  globalEnvs?: {
    all?: boolean;
  };
  ssl?: boolean;
  backup?: PackageBackup[];
  changelog?: string;
  warnings?: {
    onInstall?: string;
    onUpdate?: string;
    onReset?: string;
    onRemove?: string;
  };
  updateAlerts?: ManifestUpdateAlert[];
  disclaimer?: {
    message: string;
  };
  gettingStarted?: string;
  style?: {
    featuredBackground?: string;
    featuredColor?: string;
    featuredAvatarFilter?: string;
  };
  setupWizard?: SetupWizard;
  // Legacy setupWizardv1
  setupSchema?: SetupSchema;
  setupTarget?: SetupTarget;
  setupUiJson?: SetupUiJson;

  author?: string;
  contributors?: string[];
  categories?: string[];
  keywords?: string[];
  links?: {
    homepage?: string;
    ui?: string;
    api?: string;
    gateway?: string;
  };
  repository?: {
    type?: string;
    url?: string;
    directory?: string;
  };
  bugs?: {
    url: string;
  };
  license?: string;
}